+++ tests/test_data.py
@@ -24,6 +24,32 @@ test_message = '''
 }
 '''
 
+test_feedback_message = '''
+ {
+   "invalid": false,
+   "origin": "uk.gov.ons.edc.eq",
+   "survey_id": "feedback",
+   "survey_type": "RSI",
+   "survey_period": "2016",
+   "submitted_at": "2016-03-12T10:39:40Z",
+   "satisfaction": "satisfied",
+   "comment": "comment"
+ }
+ '''
+
+invalid_feedback_message = '''
+ {
+   "invalid": true,
+   "origin": "uk.gov.ons.edc.eq",
+   "survey_id": "feedback",
+   "survey_type": "RSI",
+   "survey_period": "2016",
+   "submitted_at": "2016-03-12T10:39:40Z",
+   "satisfaction": "satisfied",
+   "comment": "comment"
+ }
+ '''
+
 updated_message = '''
 {
   "type": "uk.gov.ons.edc.eq:surveyresponse",