+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Unreleased
+  - Add support for feedback survey responses
   - Change all instances of ADD to COPY in Dockerfile
   - Remove use of SDX_HOME variable in makefile
   - Fix Postgres DB tests